requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dashscope==1.24.7
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
distro==1.9.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fancy-einsum==0.0.3
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
iniconfig==2.3.0
ipykernel==7.0.0
ipython==8.37.0
jaxtyping==0.3.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.11.1
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kaleido==1.2.0
kiwisolver==1.4.9
lightning-utilities==0.15.2
line_profiler==5.0.0
logistro==2.0.1
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.2
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
open_clip_torch==3.2.0
openai==0.28.0
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.5.0
plotly==5.19.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
ptwt==0.1.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-lightning==2.5.5
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.41.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.22
tokenizers==0.22.1
tomli==2.3.0
torch==2.7.0
torchaudio==2.7.0
torchmetrics==1.8.2
torchvision==0.22.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
-e git+ssh://git@github.com/Prisma-Multimodal/ViT-Prisma.git@46d21f0bb1a4e23aaad60ff777e5fa3243e09383#egg=vit_prisma
wadler_lindig==0.1.7
wandb==0.22.2
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.22.0
